getty refactor
