+++ remoting/getty/client.go
@@ -1,23 +1,24 @@
 /*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+* Licensed to the Apache Software Foundation (ASF) under one or more
+* contributor license agreements.  See the NOTICE file distributed with
+* this work for additional information regarding copyright ownership.
+* The ASF licenses this file to You under the Apache License, Version 2.0
+* (the "License"); you may not use this file except in compliance with
+* the License.  You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
 
-package dubbo
+package getty
 //
 //import (
+//	"github.com/apache/dubbo-go/remoting"
 //	"math/rand"
 //	"strings"
 //	"sync"
@@ -167,62 +168,62 @@ package dubbo
 //	return c
 //}
 //
-//// Request ...
-//type Request struct {
-//	addr   string
-//	svcUrl common.URL
-//	method string
-//	args   interface{}
-//	atta   map[string]string
-//}
-//
-//// NewRequest ...
-//func NewRequest(addr string, svcUrl common.URL, method string, args interface{}, atta map[string]string) *Request {
-//	return &Request{
-//		addr:   addr,
-//		svcUrl: svcUrl,
-//		method: method,
-//		args:   args,
-//		atta:   atta,
-//	}
-//}
-//
-//// Response ...
-//type Response struct {
-//	reply interface{}
-//	atta  map[string]string
-//}
-//
-//// NewResponse ...
-//func NewResponse(reply interface{}, atta map[string]string) *Response {
-//	return &Response{
-//		reply: reply,
-//		atta:  atta,
-//	}
-//}
+////// Request ...
+////type Request struct {
+////	addr   string
+////	svcUrl common.URL
+////	method string
+////	args   interface{}
+////	atta   map[string]string
+////}
+////
+////// NewRequest ...
+////func NewRequest(addr string, svcUrl common.URL, method string, args interface{}, atta map[string]string) *Request {
+////	return &Request{
+////		addr:   addr,
+////		svcUrl: svcUrl,
+////		method: method,
+////		args:   args,
+////		atta:   atta,
+////	}
+////}
+////
+////// Response ...
+////type Response struct {
+////	reply interface{}
+////	atta  map[string]string
+////}
+////
+////// NewResponse ...
+////func NewResponse(reply interface{}, atta map[string]string) *Response {
+////	return &Response{
+////		reply: reply,
+////		atta:  atta,
+////	}
+////}
 //
 //// CallOneway call one way
-//func (c *Client) CallOneway(request *Request) error {
-//
-//	return perrors.WithStack(c.call(CT_OneWay, request, NewResponse(nil, nil), nil))
-//}
-//
-//// Call if @response is nil, the transport layer will get the response without notify the invoker.
-//func (c *Client) Call(request *Request, response *Response) error {
-//
-//	ct := CT_TwoWay
-//	if response.reply == nil {
-//		ct = CT_OneWay
-//	}
-//
-//	return perrors.WithStack(c.call(ct, request, response, nil))
-//}
-//
-//// AsyncCall ...
-//func (c *Client) AsyncCall(request *Request, callback common.AsyncCallback, response *Response) error {
-//
-//	return perrors.WithStack(c.call(CT_TwoWay, request, response, callback))
-//}
+////func (c *Client) CallOneway(request *Request) error {
+////
+////	return perrors.WithStack(c.call(CT_OneWay, request, NewResponse(nil, nil), nil))
+////}
+////
+////// Call if @response is nil, the transport layer will get the response without notify the invoker.
+////func (c *Client) Call(request *Request, response *Response) error {
+////
+////	ct := CT_TwoWay
+////	if response.reply == nil {
+////		ct = CT_OneWay
+////	}
+////
+////	return perrors.WithStack(c.call(ct, request, response, nil))
+////}
+////
+////// AsyncCall ...
+////func (c *Client) AsyncCall(request *Request, callback common.AsyncCallback, response *Response) error {
+////
+////	return perrors.WithStack(c.call(CT_TwoWay, request, response, callback))
+////}
 //
 //func (c *Client) call(ct CallType, request *Request, response *Response, callback common.AsyncCallback) error {
 //
@@ -310,57 +311,54 @@ package dubbo
 //}
 //
 //func (c *Client) heartbeat(session getty.Session) error {
-//	return c.transfer(session, nil, NewPendingResponse())
+//	req := remoting.NewRequest("2.0.2")
+//	req.TwoWay = true
+//	req.Event = true
+//	resp := remoting.NewPendingResponse()
+//	return c.transfer(session, req, 3 * time.Second, resp)
 //}
 //
-//func (c *Client) transfer(session getty.Session, pkg *DubboPackage,
-//	rsp *PendingResponse) error {
+//func (c *Client) transfer(session getty.Session, request *remoting.Request, timeout time.Duration,
+//	rsp *remoting.PendingResponse) error {
 //
-//	var (
-//		sequence uint64
-//		err      error
-//	)
-//
-//	sequence = c.sequence.Add(1)
-//
-//	if pkg == nil {
-//		pkg = &DubboPackage{}
-//		pkg.Body = hessian.NewRequest([]interface{}{}, nil)
-//		pkg.Body = []interface{}{}
-//		pkg.Header.Type = hessian.PackageHeartbeat
-//		pkg.Header.SerialID = byte(S_Dubbo)
-//	}
-//	pkg.Header.ID = int64(sequence)
+//	//sequence = c.sequence.Add(1)
+//	//
+//	//if pkg == nil {
+//	//	pkg = &DubboPackage{}
+//	//	pkg.Body = hessian.NewRequest([]interface{}{}, nil)
+//	//	pkg.Body = []interface{}{}
+//	//	pkg.Header.Type = hessian.PackageHeartbeat
+//	//	pkg.Header.SerialID = byte(S_Dubbo)
+//	//}
+//	//pkg.Header.ID = int64(sequence)
 //
 //	// cond1
-//	if rsp != nil {
-//		rsp.seq = sequence
-//		c.addPendingResponse(rsp)
-//	}
+//	//if rsp != nil {
+//	//	c.addPendingResponse(rsp)
+//	//}
 //
-//	err = session.WritePkg(pkg, c.opts.RequestTimeout)
-//	if err != nil {
-//		c.removePendingResponse(SequenceType(rsp.seq))
-//	} else if rsp != nil { // cond2
+//	err := session.WritePkg(request, timeout)
+//	if rsp != nil { // cond2
 //		// cond2 should not merged with cond1. cause the response package may be returned very
 //		// soon and it will be handled by other goroutine.
-//		rsp.readStart = time.Now()
+//		rsp.ReadStart = time.Now()
 //	}
 //
 //	return perrors.WithStack(err)
 //}
 //
-//func (c *Client) addPendingResponse(pr *PendingResponse) {
-//	c.pendingResponses.Store(SequenceType(pr.seq), pr)
-//}
-//
-//func (c *Client) removePendingResponse(seq SequenceType) *PendingResponse {
-//	if c.pendingResponses == nil {
-//		return nil
-//	}
-//	if presp, ok := c.pendingResponses.Load(seq); ok {
-//		c.pendingResponses.Delete(seq)
-//		return presp.(*PendingResponse)
-//	}
-//	return nil
-//}
+////
+////func (c *Client) addPendingResponse(pr *PendingResponse) {
+////	c.pendingResponses.Store(SequenceType(pr.seq), pr)
+////}
+////
+////func (c *Client) removePendingResponse(seq SequenceType) *PendingResponse {
+////	if c.pendingResponses == nil {
+////		return nil
+////	}
+////	if presp, ok := c.pendingResponses.Load(seq); ok {
+////		c.pendingResponses.Delete(seq)
+////		return presp.(*PendingResponse)
+////	}
+////	return nil
+////}
