+++ remoting/getty/client_test.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package dubbo
+package getty
 //
 //import (
 //	"bytes"
